submit Bill BillDetail

+++ src/main/java/com/example/lotte/service/CustomerService.java
@@ -1,8 +1,12 @@
 package com.example.lotte.service;
 
 import com.example.lotte.DTO.CustomerDTO;
+import com.example.lotte.DTO.RankDTO;
 import com.example.lotte.model.Customer;
+import com.example.lotte.model.Rank;
+import com.example.lotte.repository.BillRepository;
 import com.example.lotte.repository.CustomerRepository;
+import com.example.lotte.repository.RankRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -16,6 +20,11 @@ public class CustomerService {
 
     @Autowired
     CustomerRepository customerRepository;
+    @Autowired
+    RankRepository rankRepository;
+
+    @Autowired
+    BillRepository billRepository;
 
     public Customer addCustomer(CustomerDTO customerDTO) {
         Customer customer = new Customer();
@@ -24,6 +33,8 @@ public class CustomerService {
         customer.setDob(customerDTO.getDob());
         customer.setAddress(customerDTO.getAddress());
         customer.setPhoneNumber(customerDTO.getPhoneNumber());
+        Rank rank = rankRepository.findById(Long.valueOf(4)).get();
+        customer.setRank(rank);
         customer.setTotalPoint(0);
 
         Customer savedCustomer = customerRepository.save(customer);
@@ -55,7 +66,7 @@ public class CustomerService {
             customerRepository.save(customer);
             return ResponseEntity.ok("Customer updated successfully.");
         } else {
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.ok("Customer id not found");
         }
     }
 
@@ -64,17 +75,30 @@ public class CustomerService {
 
         if (optionalCustomer.isPresent()) {
             Customer customer = optionalCustomer.get();
-            customerRepository.delete(customer);
-            return ResponseEntity.ok("Customer deleted successfully.");
+            if(billRepository.findAllByCustomerId(customer.getId()).size() > 0) {
+                return ResponseEntity.ok("Customer  has  a bill ");
+            }
+            else {
+                customerRepository.delete(customer);
+                return ResponseEntity.ok("Customer deleted successfully.");
+            }
         } else {
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.ok("Customer id not found");
         }
     }
 
     public List<CustomerDTO> getAllCustomer() {
         List<Customer> customers = customerRepository.findAll();
+
         List<CustomerDTO> customerDTOs = customers.stream()
                 .map(customer -> {
+                    RankDTO rankDTO = new RankDTO();
+                    if (customer.getRank() != null ) {
+                        Optional<Rank> rank = rankRepository.findById(customer.getRank().getId());
+                        if(rank.isPresent()) {
+                            rankDTO = new RankDTO(rank.get().getId(), rank.get().getName());
+                        }
+                    }
                     CustomerDTO dto = new CustomerDTO();
                     dto.setId(customer.getId());
                     dto.setName(customer.getName());
@@ -82,6 +106,8 @@ public class CustomerService {
                     dto.setDob(customer.getDob());
                     dto.setAddress(customer.getAddress());
                     dto.setPhoneNumber(customer.getPhoneNumber());
+                    dto.setTotalPoint(customer.getTotalPoint());
+                    dto.setRank(rankDTO);
                     return dto;
                 })
                 .collect(Collectors.toList());
